refactor: acceptance 패키지 세분화하여 정리
